Fehlermeldungen hinzugefügt
Nur ein Upload von Bilddatein möglich +Fehlermeldungen

+++ webapp/pictureupload.py
@@ -50,6 +50,44 @@ def writehtml(info):
 </html>
 """ % (info['product_id'],info['product_name'],info['inhalte'],info['sonstiges'],info['bild'])
 
+def writeerror(info):
+    print 'Content-Type: text/html\n\n'
+    reload(sys)
+    sys.setdefaultencoding('utf-8')
+    print """\
+<!DOCTYPE HTML>
+<html>
+        <head>
+                <title>Fehler:</title>
+                <meta charset="utf-8" />
+                <meta name="viewport" content="width=device-width, initial-scale=1" />
+                <link rel="stylesheet" href="/assets/css/main.css" />
+                <noscript><link rel="stylesheet" href="/assets/css/noscript.css" /></noscript>
+        </head>
+        <body>
+                <!-- Wrapper -->
+                        <div id="wrapper">
+                                <!-- Main -->
+                                        <section id="main">
+                                                <header>
+                                                        <span class="avatar"><img src="/images/barcode_scanner.png" alt="" /></span>
+                                                        </header>
+							<h1>Fehler:</h1>
+                                                        <br><p>%s</p>
+                                                </footer>
+                                        </section>
+                                <!-- Footer -->
+                                        <footer id="footer">
+                                                <ul class="copyright">
+                                                        <li>&copy; by Team IssEss</li>
+                                                        <li>Design: <a href="http://twitter.com/3zscan">by Foz</a></li>
+                                                </ul>
+                                        </footer>
+                        </div>
+        </body>
+</html>
+""" % (info['art'])
+
 def writejson(info):
     print 'Content-Type: application/json\n\n'
     reload(sys)
@@ -83,7 +121,7 @@ def decode(f):
 
     # clean up
     del(image)
-    return ".4029764001807"
+    return "---"
     
     
 
@@ -96,31 +134,81 @@ form = cgi.FieldStorage()
 if form.getvalue('userfile'):
     fileitem=form['userfile']
     #yay...we got a file
+    ean="nopicture"
     fn = '/tmp/' + os.path.basename(fileitem.filename)
-    open(fn, 'wb').write(fileitem.file.read())
-    ean=decode(fn)
-    os.remove(fn)
+
+    if ".png" in fn:
+    	open(fn, 'wb').write(fileitem.file.read())
+    	ean=decode(fn)
+        os.remove(fn)
+
+    if ".jpg" in fn:
+	open(fn, 'wb').write(fileitem.file.read())
+        ean=decode(fn)
+        os.remove(fn)
+
+    if ".JPG" in fn:
+        open(fn, 'wb').write(fileitem.file.read())
+        ean=decode(fn)
+        os.remove(fn)
 
 if form.getvalue('ean'):
     ean=str(form.getvalue('ean'))
-    
+
+if ean == "nopicture":
+    info={}
+    info['art']="Bitte wähle ein Foto aus!"
+    writeerror(info)
+    exit()
+
+if ean == "---":
+    info={}
+    info['art']="Scan fehlgeschlagen, das Foto ist nicht lesbar."
+    writeerror(info)
+    exit()
+	   
 data = json.load(urllib2.urlopen('http://world.openfoodfacts.org/api/v0/product/' + ean +'.json'))
-info={}
-info['product_id']=data['product']['code']
-info['inhalte']=data['product']['ingredients_text']
-info['product_name']=data['product']['product_name']
-info['bild']=data['product']['image_front_url']
-info['sonstiges']=data['product']['labels_hierarchy']
+find = """/
+%s
+""" % (data)
+find = find.replace(" ", "-")
+if "product-found" in find:
+		info={}
+		info['product_id']=data['product']['code']
+		info['inhalte']=data['product']['ingredients_text']
+		info['product_name']=data['product']['product_name']
+		info['bild']=data['product']['image_front_url']
+		info['sonstiges']='Kommt bald...'#fn#data['product']['labels_hierarchy']
 
-usetztung = csv.reader(open('Uebersetzungstabelle2.csv','rb'))
-for row in usetztung:
-    alt=row[0].decode('utf-8')
-    neu=row[1].decode('utf-8')
-    info['inhalte']=info['inhalte'].replace(alt, neu)
-    
+		usetztung = csv.reader(open('Uebersetzungstabelle2.csv','rb'))
+		for row in usetztung:
+    			alt=row[0].decode('utf-8')
+    			neu=row[1].decode('utf-8')
+    			info['inhalte']=info['inhalte'].replace(alt, neu)
+		if form.getvalue('as_json'):
+                    writejson(info)
+                    exit()
+        	else:
+                    writehtml(info)
+                    exit()    
 
-
-if form.getvalue('as_json'):
-    writejson(info)
+if "'product-not-found'" in find:
+	info={}
+	info['art']="Product-ID:<br>" + ean + "<br>Dieses Produkt existiert noch nicht in der Datenbank."
+	info['product_id']=ean
+	if form.getvalue('as_json'):
+            writejson(info)
+            exit()
+        else:
+            writeerror(info)
+            exit()
 else:
-    writehtml(info)
+	info={}
+	info['art']="Ein Unbekannter Fehler ist aufgetreten!"
+
+	if form.getvalue('as_json'):
+    	    writejson(info)
+            exit()
+	else:
+	    writeerror(info)
+            exit()
